issue-12 fixed registration component to work with laravel octane
